requirements/common.txt
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
Faker==12.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
iniconfig==2.0.0
mixer==7.2.2
packaging==24.1
pluggy==1.5.0
pytest==8.2.2
python-dateutil==2.9.0.post0
python-telegram-bot==21.4
requests==2.32.3
setuptools==70.3.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
urllib3==2.2.2
wheel==0.43.0
